--- packages/app/src/domain/escalation-level/scoreboard/logic/index.tsx
@@ -1,2 +1,0 @@
-export * from './scoreboard-types';
-export * from './sort-options';
