+++ packages/app/src/domain/escalation-level/scoreboard/types.d.ts
