+++ packages/app/src/domain/variants/variants-table-tile/components/index.ts
@@ -1,0 +1,7 @@
+export * from './narrow-variants-table';
+export * from './percentage-bar-with-number';
+export * from './shared-table-components';
+export * from './variant-difference';
+export * from './variant-name-cell';
+export * from './variants-table';
+export * from './wide-variants-table';
