--- packages/app/src/domain/variants/variants-table-tile/components/index.tsx
@@ -1,6 +1,0 @@
-export { NarrowVariantsTable } from './narrow-variants-table';
-export { PercentageBarWithNumber } from './percentage-bar-with-number';
-export { Cell, HeaderCell, StyledTable } from './shared-table-components';
-export { VariantDifference } from './variant-difference';
-export { VariantNameCell } from './variant-name-cell';
-export { WideVariantsTable } from './wide-variants-table';
