+++ packages/app/src/domain/variants/variants-table-tile/components/narrow-variants-table.tsx
@@ -5,7 +5,6 @@ import styled from 'styled-components';
 import { isPresent } from 'ts-is-present';
 import { Box } from '~/components/base';
 import { InlineText } from '~/components/typography';
-import { TableText } from '~/domain/variants/variants-table-tile';
 import { useIntl } from '~/intl';
 import { VariantRow } from '~/static-props/variants/get-variant-table-data';
 import { getMaximumNumberOfDecimals } from '~/utils/get-maximum-number-of-decimals';
@@ -19,6 +18,7 @@ import {
   VariantNameCell,
 } from '.';
 import { useVariantNameAndDescription } from '../logic/use-variant-name-and-description';
+import { TableText } from '../types';
 import { NoPercentageData } from './no-percentage-data';
 
 type NarrowVariantsTableProps = {
