+++ packages/app/src/domain/variants/variants-table-tile/index.ts
@@ -1,0 +1,1 @@
+export * from './variants-table-tile';
