+++ packages/app/src/domain/variants/variants-table-tile/logic/index.ts
@@ -1,0 +1,1 @@
+export * from './use-variant-name-and-description';
