+++ packages/app/src/domain/variants/variants-table-tile/types.d.ts
@@ -1,0 +1,14 @@
+export type TableText = {
+  anderen_tooltip: string;
+  omschrijving: string;
+  omschrijving_zonder_placeholders: string;
+  titel: string;
+  kolommen: {
+    aantal_monsters: string;
+    eerst_gevonden: string;
+    percentage: string;
+    variant_titel: string;
+    vorige_meting: string;
+  };
+  verschil: { gelijk: string; meer: string; minder: string };
+};
