+++ packages/app/src/domain/variants/variants-table-tile/variants-table-tile.tsx
@@ -12,21 +12,7 @@ import { useIntl } from '~/intl';
 import { VariantRow } from '~/static-props/variants/get-variant-table-data';
 import { replaceVariablesInText } from '~/utils/replace-variables-in-text';
 import { VariantsTable } from './components/variants-table';
-
-export type TableText = {
-  anderen_tooltip: string;
-  omschrijving: string;
-  omschrijving_zonder_placeholders: string;
-  titel: string;
-  kolommen: {
-    aantal_monsters: string;
-    eerst_gevonden: string;
-    percentage: string;
-    variant_titel: string;
-    vorige_meting: string;
-  };
-  verschil: { gelijk: string; meer: string; minder: string };
-};
+import { TableText } from './types';
 
 export function VariantsTableTile({
   text,
