+++ packages/app/src/pages/over-risiconiveaus.tsx
@@ -11,8 +11,10 @@ import { VrChoropleth } from '~/components/choropleth/vr-choropleth';
 import { RichContent } from '~/components/cms/rich-content';
 import { ErrorBoundary } from '~/components/error-boundary';
 import { Heading, InlineText, Text } from '~/components/typography';
-import { Scoreboard } from '~/domain/escalation-level/scoreboard';
-import { selectScoreboardData } from '~/domain/escalation-level/scoreboard/data-selection/select-scoreboard-data';
+import {
+  Scoreboard,
+  selectScoreboardData,
+} from '~/domain/escalation-level/scoreboard';
 import { Layout } from '~/domain/layout/layout';
 import { useIntl } from '~/intl';
 import {
